+++ src/Resources/Subscriptions/Addons.ts
@@ -1,8 +1,8 @@
 import AbstractResource from '../AbstractResource';
 
-export default class Subscriptions extends AbstractResource {
+export default class Addons extends AbstractResource {
     initialise(): void {
-        this.resourcePath = '/subscriptions';
+        this.resourcePath = '/subscription-addons';
         this.defaultAccess = 'guest';
     }
